data lock events moved to Events namespace
