+++ src/SFA.DAS.Payments.AcceptanceTests.EndToEnd/Handlers/PayableEarningEventHandler.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Concurrent;
 using System.Threading.Tasks;
 using NServiceBus;
-using SFA.DAS.Payments.DataLocks.Messages;
+using SFA.DAS.Payments.DataLocks.Messages.Events;
 
 namespace SFA.DAS.Payments.AcceptanceTests.EndToEnd.Handlers
 {
