+++ src/SFA.DAS.Payments.DataLocks.Application/Mapping/DataLocksProfile.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using SFA.DAS.Payments.DataLocks.Messages;
+using SFA.DAS.Payments.DataLocks.Messages.Events;
 using SFA.DAS.Payments.EarningEvents.Messages.Events;
 
 namespace SFA.DAS.Payments.DataLocks.Application.Mapping
