+++ src/SFA.DAS.Payments.Datalocks.Messages/Events/PayableEarningEvent.cs
@@ -1,4 +1,4 @@
-﻿namespace SFA.DAS.Payments.DataLocks.Messages
+﻿namespace SFA.DAS.Payments.DataLocks.Messages.Events
 {
     public class PayableEarningEvent : DataLockEvent
     {
