+++ src/SFA.DAS.Payments.RequiredPayments.Application/Processors/PayableEarningEventProcessor.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using SFA.DAS.Payments.DataLocks.Messages;
+using SFA.DAS.Payments.DataLocks.Messages.Events;
 using SFA.DAS.Payments.RequiredPayments.Domain;
 using SFA.DAS.Payments.RequiredPayments.Messages.Events;
